# api/openapi.yaml
openapi: 3.0.3
...
info:
  title: microvm
  version: "1.0.0"
  description: |
    Every sandbox is a real virtual machine with its own kernel. Code inside one
    is assumed hostile.

    There are two ways to run code, and the difference is the point:

    - **Sandboxes** are held. You create one, run commands in it, and it is
      yours until you stop it or its TTL elapses. Creation fails when the node
      is full, so backpressure is yours to handle.
    - **Tasks** are queued. You hand work to the fleet and it runs when a slot
      frees anywhere. Submission never fails for capacity.

    Use a sandbox for several commands sharing state; use a task for throughput.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080/v1
    description: A local daemon. Bound to loopback by default, because this API creates VMs that run arbitrary code.
security:
  - bearerAuth: []
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      summary: List tenant policies
      description: |
        Every tenant with a storage policy set. **Admin only.** A tenant with no
        policy is unlimited and does not appear here.
      tags: [Tenants]
      responses:
        '200':
          description: The configured tenants.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
  /tenants/{tenant}:
    parameters:
      - name: tenant
        in: path
        required: true
        schema: { type: string }
        description: The tenant namespace, e.g. from a sandbox's `storage.namespace`.
    put:
      operationId: updateTenant
      summary: Set a tenant's storage policy
      description: |
        Set the storage cap and full-policy for a tenant. **Admin only.** This is
        the one thing an admin key can do that a normal key cannot, and it is
        powerful: `evict` deletes a tenant's oldest data to make room, so setting
        it is a decision to allow that deletion.
      tags: [Tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantUpdateParams' }
      responses:
        '200':
          description: The updated tenant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
    get:
      operationId: retrieveTenant
      summary: Retrieve a tenant's policy and usage
      description: |
        A tenant's cap, policy, and how much of the cap is used right now.
        **Admin only.** The usage is read from the object store, so it is the
        real total, not an estimate.
      tags: [Tenants]
      responses:
        '200':
          description: The tenant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>`. Tokens are compared in constant time.
...
  responses:
    InvalidRequest:
      description: The request is malformed or contradicts itself.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: The token is missing or wrong.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: |
        The token is valid but lacks the power for this. Distinct from 401 on
        purpose: 401 means "log in", 403 means "you are logged in and still may
        not". Setting a tenant policy needs an admin key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: No such object.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    CapacityExhausted:
      description: |
        No room right now. Either the node has no free slot, or the caller is over
        a per-tenant limit -- both are `capacity_error` with code
        `node_at_capacity`, deliberately, because the caller's move is the same and
        that is the pair every SDK already backs off on. The message is what tells
        them apart: a full node frees up on its own, whereas being over your own
        sandbox cap needs you to delete one.
      headers:
        Retry-After:
          description: |
            Seconds to wait, on the refusals whose end can be named to the second --
            the request-rate limit. Absent for a full node and for the per-tenant
            sandbox cap: when a slot frees is not something this node knows, and a
            number invented here would be a guess dressed as a fact.
          required: false
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      description: The body of every non-2xx reply.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/Error' }
    Error:
      type: object
      required: [type, code, message]
      properties:
        type:
          $ref: '#/components/schemas/ErrorType'
        code:
          type: string
          description: |
            A stable, machine-readable identifier. Switch on this. The message is
            for humans and may be reworded at any time.
          example: "sandbox_not_found"
        message:
          type: string
          description: A human-readable explanation. Never parse it.
          example: "No such sandbox: sb_01JZ8QK3M4N5P6R7S8T9V0W1X2"
        param:
          type: string
          description: Which request field was at fault, when one was.
          example: "image"
        request_id:
          type: string
          description: Quote this when reporting a problem.
    ErrorType:
      type: string
      description: |
        The class of error, which determines what a client should *do*:

        - `invalid_request_error` — the request is wrong. Retrying it unchanged
          will fail identically.
        - `authentication_error` — the token is missing or wrong.
        - `capacity_error` — no room right now. Retrying later may work; a task
          would not have failed at all.
        - `idempotency_error` — an Idempotency-Key was reused with a different
          body.
        - `api_error` — our fault. Safe to retry with the same Idempotency-Key.
      enum:
        - invalid_request_error
        - authentication_error
        - capacity_error
        - idempotency_error
        - api_error

    # --- pagination ---------------------------------------------------------
    ListEnvelope:
      type: object
      required: [object, url, has_more, data]
      properties:
        object: { type: string, enum: [list] }
        url:
          type: string
          description: The endpoint this list came from.
        has_more:
          type: boolean
          description: |
            Whether more objects exist past this page. Page on with
            `starting_after` set to the last ID in `data`.
        data:
          type: array
          items:
            type: object

    # --- sandbox ------------------------------------------------------------
...
    TenantFullPolicy:
      type: string
      description: |
        What a write does when a tenant is at its cap.

        - `preserve` — reject the write and keep what is stored. For data that
          matters more than the next write: an audit trail, a paid-for dataset.
        - `evict` — delete the oldest objects until the write fits. For a cache
          or scratch space, where the newest data is the point. It deletes data,
          irreversibly, which is why it is never the default.
      enum: [preserve, evict]
    Tenant:
      type: object
      required: [object, id, max_bytes, policy]
      properties:
        object: { type: string, enum: [tenant] }
        id:
          type: string
          description: The tenant namespace this policy applies to.
          example: "t_9f86d081884c7d659a2feaa0c55ad015"
        max_bytes:
          type: integer
          format: int64
          description: The tenant's total storage cap, in bytes. 0 means unlimited.
        policy: { $ref: '#/components/schemas/TenantFullPolicy' }
        usage_bytes:
          type: integer
          format: int64
          description: |
            Bytes currently stored, read from the object store on retrieve. Absent
            on the reply to an update, which does not stop to scan the bucket.
    TenantList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Tenant' }
    TenantUpdateParams:
      type: object
      required: [max_bytes, policy]
      properties:
        max_bytes:
          type: integer
          format: int64
          description: The tenant's total storage cap, in bytes. 0 means unlimited.
          example: 1073741824
        policy: { $ref: '#/components/schemas/TenantFullPolicy' }

    # --- execution ----------------------------------------------------------
...